[#MHV-39899] unread message count on dashboard
