+++ src/applications/mhv/secure-messaging/util/helpers.js
@@ -31,6 +31,12 @@ export const navigateToFolderByFolderId = (folderId, history) => {
   history.push(folderPathByFolderId(folderId));
 };
 
+export const unreadCountAllFolders = folders => {
+  return folders
+    .filter(folder => folder.id !== Folders.DRAFTS.id)
+    .reduce((a, b) => a + b.unreadCount, 0);
+};
+
 export const navigateToFoldersPage = history => {
   history.push('/folders');
 };
